Move the StandardPlugin into a "standard" package along with the ops.
