+++ core/src/main/kotlin/org/nwapw/abacus/plugin/standard/operator/OperatorNcr.kt
@@ -5,7 +5,7 @@ import org.nwapw.abacus.function.NumberOperator
 import org.nwapw.abacus.function.OperatorAssociativity
 import org.nwapw.abacus.function.OperatorType
 import org.nwapw.abacus.number.NumberInterface
-import org.nwapw.abacus.plugin.StandardPlugin.*
+import org.nwapw.abacus.plugin.standard.StandardPlugin.*
 
 /**
  * The "N choose R" operator.
